:sparkles: 프로필 사진 삭제 예외처리

+++ src/main/java/Idea/Archive/IdeaArchive/domain/img/service/ChangeProfileService.java
@@ -1,6 +1,8 @@
 package Idea.Archive.IdeaArchive.domain.img.service;
 
+import Idea.Archive.IdeaArchive.domain.img.exception.NotExistImageException;
 import Idea.Archive.IdeaArchive.domain.member.entity.Member;
+import Idea.Archive.IdeaArchive.global.img.DefaultImage;
 import Idea.Archive.IdeaArchive.global.util.MemberUtil;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
@@ -20,6 +22,7 @@ public class ChangeProfileService {
     private final UploadProfileImg uploadProfileImg;
     private final AmazonS3 amazonS3;
     private final MemberUtil memberUtil;
+    private final DefaultImage defaultImage;
 
     private void deleteImage(String fileName) {
         amazonS3.deleteObject(new DeleteObjectRequest(bucket, fileName));
@@ -27,6 +30,9 @@ public class ChangeProfileService {
 
     public void execute(List<MultipartFile> multipartFileList) {
         Member currentMember = memberUtil.currentMember();
+        if (currentMember.getProfileImageUrl().equals(defaultImage)) {
+            throw new NotExistImageException("기본 프로필은 삭제할 수 없습니다.");
+        }
         deleteImage(currentMember.getProfileImageUrl());
         uploadProfileImg.execute(multipartFileList);
     }
